Create update index.html
Tried to delete the last one, looks like it did but no sure.  I don't want so many commits as I work out the kinks.  It all works, please commit if we are ok with this layout so that we can start the css and javascript.  It looks so basic but I think we can spice it up right nice.  I can't wait to see what it looks like.